Split out the MapController into submodules
