+++ app/views/layouts/application.html.erb
@@ -38,15 +38,6 @@
       <%= yield %>
     </main>
 
-    <!-- Hidden modal container to be populated dynamically by Stimulus controllers -->
-    <div id="modal"
-      role="dialog"
-      aria-modal="true"
-      class="fixed top-0 right-0 h-screen w-full md:w-1/3 md:max-w-xl bg-white shadow-2xl p-8 md:rounded-bl-lg transition-width duration-300 overflow-y-auto max-h-screen"
-      style="display:none; z-index:2000;"></div>
-
-
-
     <%# JS template for the house edit form %>
     <script type="text/template" id="house-edit-form-template">
     <div class="house-details bg-white max-w-md mx-auto">
