+++ app/views/pages/map.html.erb
@@ -13,4 +13,11 @@
 
   <!-- Map Container -->
   <div data-map-target="canvas" class="map-fullscreen"></div>
+
+  <!-- Modal Container -->
+  <div data-map-target="modal"
+    role="dialog"
+    aria-modal="true"
+    class="fixed top-0 right-0 h-screen w-full md:w-1/3 md:max-w-xl bg-white shadow-2xl p-8 md:rounded-bl-lg transition-width duration-300 overflow-y-auto max-h-screen"
+    style="display:none; z-index:2000;"></div>
 </div>
